Fix Relaxer aggregation on subsequent tests
