+++ examples/src/commonTest/kotlin/tech/antibytes/kmock/example/SampleControllerAutoStubRelaxedSpec.kt
@@ -11,12 +11,10 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import tech.antibytes.kfixture.PublicApi
 import tech.antibytes.kfixture.fixture
 import tech.antibytes.kfixture.kotlinFixture
 import tech.antibytes.kfixture.listFixture
 import tech.antibytes.kmock.MockCommon
-import tech.antibytes.kmock.Relaxer
 import tech.antibytes.kmock.example.contract.ExampleContract
 import tech.antibytes.kmock.example.contract.ExampleContract.SampleDomainObject
 import tech.antibytes.kmock.example.contract.ExampleContract.SampleLocalRepository
@@ -36,35 +34,8 @@ import tech.antibytes.util.test.coroutine.runBlockingTestWithTimeoutInScope
 import tech.antibytes.util.test.fulfils
 import tech.antibytes.util.test.mustBe
 import kotlin.js.JsName
-import kotlin.native.concurrent.ThreadLocal
-import kotlin.reflect.KClass
 import kotlin.test.BeforeTest
 import kotlin.test.Test
-
-@ThreadLocal
-object Fixture {
-    var fixture: PublicApi.Fixture? = null
-}
-
-@Relaxer
-@Suppress("UNUSED_PARAMETER")
-internal inline fun <reified T> relax(id: String): T {
-    if (Fixture.fixture == null) {
-        Fixture.fixture = kotlinFixture()
-    }
-
-    return Fixture.fixture!!.fixture()
-}
-
-@Suppress("UNCHECKED_CAST", "UNUSED_PARAMETER")
-internal fun <T> relax(id: String, type0: KClass<CharSequence>, type1: KClass<Comparable<*>>): T {
-    return Fixture.fixture!!.fixture<String>() as T
-}
-
-@Suppress("UNCHECKED_CAST", "UNUSED_PARAMETER")
-internal fun <T> relax(id: String, type0: KClass<Any>): T {
-    return Fixture.fixture!!.fixture<Int>() as T
-}
 
 @MockCommon(
     SampleRemoteRepository::class,
